AV-16 #comment Refactored ProductionCycle class
